requirements.txt
asgiref==3.4.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.13
django-filter==21.1
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==3.2.0
Markdown==3.3.7
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
psycopg2==2.9.3
PyJWT==1.7.1
pyparsing==3.0.7
pyrsistent==0.18.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==5.3.0
zipp==3.6.0
